@@ -2,8 +2,8 @@ package com.stypox.mastercom_workbook.view.holder;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.HorizontalScrollView;
 import android.widget.TextView;
