@@ -1,6 +1,6 @@
 package com.stypox.mastercom_workbook.view.holder;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.View;
 
 public interface ItemHolderFactory<D> {
